Added AbstractJDBCDAO, To be refactored
